requirements.txt
Django==4.2.1
psycopg2-binary==2.9.6
djangorestframework==3.14.0
django-filter==23.2
celery[redis]==5.3.0
flower==2.0.0
requests==2.31.0
